chore: js to ts + create seeker class + postinstall script
